Refactoring core route builders to allow for future extensibility
